Fixed wget+github error.
For some reason wget would report a connection error when trying
to connect to github to download Z3. The error was hard to reproduce
since it was not being reported always, but it would certainly
appear when vagrant would run build.sh. Adding GET as a request
type seems to help for some reason.