Fixed a bug related to splitting memory based on non-overlapping offsets
The issue was that when determining whether two pointers overlap based
on their offsets and sizes, their sizes where not being computed at all,
and hence a default values was used (which was unsigned max). As a
consequence, SMACK would never split memory based on non-overlapping
offsets. This is now fixed as sizes are being properly computed using
DataLayout. In turn, dynamically allocated memory if being split based
on offsets. We still have to deal with globals somehow, which are
not being split.