Resolved a non-terminating recursion issue within DSA
The DSA's DSGraph::computeNodeMapping method did not handle a particular
corner case properly, which would lead to non-terminating recursion.
This would happen in rare cases when a node would point back to itself,
to an offset that is smaller than the originating offset. Recursively
going back to smaller offsets of the same node would in turn cause
same fields/links to be analyzed again, including the one we took
to get to the smaller offset in the first place. This resulted in
non-terminating recursion.
I fixed this by guarding against recursing in such situations:
computeNodeMapping now does not recurse to smaller offsets of the same
node.

Fixes #127