Initializing DSMonitor field in GraphBuilder constructor
GraphBuilder has two constructors, and in one of them its DSMonitor
field was not being initialized. The constructor in question gets
invoked for construction of the globals graph. As it turns out,
in some situations methods of an uninitialized DSMonitor would be
invoked even during the globals graph construction, which would
result in segmentation fault. I fixed this by initializing DSMonitor
in both constructors.