Fixed a bug related to my misunderstanding of LLVM's loop unroll pass
It appears that UnrollCount argument of LLVM's loop unroll pass
indicates how many times a loop should be unrolled, no matter what
its loop count is. Hence, there was a mismatch between how this
argument was used within SMACK and what actually happens in the
LLVM's loop unroll pass. Now static-unroll is a binary argument
and it only turns of LLVM's default loop unroll settings, which
seem to be good for our purpose.