Fix to svcomp mode's filters.py.
Adds additional filter to avoid redefinition error during clang
compilation of pthread programs.

These are necessary because svcomp passes preprocessed files to SMACK,
so the pthread includes are pre-included, using the versions of pthreads
from the svcomp benchmark maintainer's machine.  This clashes with our
defintions for pthread modules, so we have to filter these lines out of
input files.