Devirtualization bug fix related to caching of bounce functions
This was a crazy bug. Devirtualization maintains a cache of bounce
functions. Whenever a new bounce function is needed, we first search
the cache to see if we have a matching function there. There was
a tricky corner case where vararg functions would break this matching.
Basically, the problem was that types of two vararg functions would
match, while their actual arguments that those functions are being
invoked with would not. And matching in the cache was only done
based on function type, which was not sufficient for this corner
case. I extended matching in the cache to match arguments one by one
as well.