Disabled DSA type inference optimization
I noticed that type inference optimization code in DSA was crashing on
some larger benchmarks. Further inspection revealed that parts of
this code look really sketchy and even do not make much sense.
As I learned from John Criswell, type inference optimization code
in DSA is hacky and it has not been thoroughly tested. Hence, I
explored potential consequences of disabling that optimization.
We run all of the SVCOMP benchmarks with and without this optimization.
The results show that there is basically no scalability degradation
when the optimization is turned off, while that fixes several crashes.
Hence, I decided to disable the optimization.

Closes #178