Added memory leak safety checks under memory-safety flag
Added global memory counter whose initiated to 0. Every malloc call
increases it and free call decreases it. Inserting
__SMACK_check_memory_leak at the end of the main function which asserts
that global memory count is 0. In free function also checking if the
parameter is not NULL and also checking if the parameter is the base
address (to make sure that free is not being called on partial memory)