Remove hack for depreciated __VERIFIER_atomic_* func
Svcomp has depreciated the convention that functions named like
__VERIFIER_atomic_*() should be run atomically.  Last year, there were
functions using this depreciated convention, so we needed a hack to
evaluate such functions as if they were being run atomically.

Our solution for this was to add the non-depreciated
__VERIFIER_atomic_begin() and __VERIFIER_atomic_end() before and after,
respectively, any call to a __VERIFIER_atomic*() call.

After last year's competition, we removed any use of such deprecaited
calls from the sv-benchmarks repo, avoiding the need for this hack.

This commit removes this hack.