Fix bugs in error witness generation script
First, do not generate error witnesses for concurrent programs because
1) it is not required per SVCOMP rules. 2) the assumption scope is
obtained by call stack whereas calls and returns may not match for
concurrent programs. Second, add warnings for calls to function pointer
dispatch procedures and raise a runtime error when call stack is broken.

TODO: fix the bug when the there is function pointer dispatch procedure
in the error trace. There is no source location assumption for calls
in function pointer dispatch procedure so there may not be call to dispatched
function in the error trace, which in turn breaks witness generation.