Updated and cleaned up our DSA wrapper class
As it turns out, we have not really been using DSAAliasAnalysis as an LLVM
alias analysis pass for a while now. Instead, DSAAliasAnalysis has been used
mainly as a SMACK wrapper around DSA. Hence, I removed the alias function and
its helper functions, which we were not invoking anyways in the first place.