Optimized inlining of alloc procedures
Our `$alloc` procedure used to be invoked both from `$malloc` and
from source code directly. This was not ideal since it makes sense
to inline it directly into `$malloc`, while in most cases it does not make
sense to inline it into source code ahead of time (from the perspective
of performance).
Now I added a level of indirection by creating `$$alloc` that both
`$malloc` and `$alloc` invoke, and `$$alloc` is being inlined
ahead of time.