Force merging of DSA nodes of pthread return values
Both pthread_join and pthread_exit manipulate return values of pthreads,
which are typically stored in memory. Given that modeling is done using
Boogie code, DSA cannot see that the return value pointers in these two
functions should alias. This is a hacky fix that addresses this issue
by introducing a dummy global variables that forces the nodes to be
merged.