Added assumption to alloc that size is not negative
For some reason when size was negative we used to allocate an
object of size 1. This seems like a hack to me and I am not sure
why it was introduced in the first place. I cleaned this up now
by requiring that size is not negative, and blocking executions
with negative size.