Fixed a bug in how I used free requires
Of course, free requires clause is not added as an assumption
at the call site! I was dumb to think that. What I needed is
to use ensures instead of free requires. This is now fixed.
Also, I switched to sticking with sle and slt instead of
also mixing it gle and glt.