Fixed memory models when flag --bit-precise-pointers is enabled
When pointers are modeled using bit-vectors, current memory
allocation procedure can return aliasing pointers because when
signed bit-vector addition overflows, the result is a negative
number which satisfies the constraint in the procedure. I fixed
it by replacing the addition with a semantically equivalent
subtraction under unbounded integers such that overflow is avoided.

Fixes #232.