Added a constraint to memory model NO_REUSE
We restrict the pointers returned by $$alloc to be less than $MALLOC_TOP,
which is defined as INT_MAX. In this way, when such pointers are stored to
and then loaded from memory when the flag --bit-precise is enabled, we get
the exact pointers back. Otherwise, the retrieved pointers can be arbitrary
because int2bv seems uninterpreted when the input argument is out-of-bound.

Partially fixes #230 in conjunction with PR #257