Fixed some issues with the modifies pass and smack.c causing the memory-safety regtests to fail.
The modifies pass was not handling the prelude section of Boogie files.
When the --memory-safety option is specified, this prelude section
contains a procedure called $global_allocations which modifies some
variables. Additionally, the modifies pass only adds modifies
clauses to SMACK-generated procedures that do not already have one, which
does not cover the $malloc procedure when the --memory-safety option is
set. This is directly changed in smack.c.