Update C string test unroll bound
The boogie verifier was not verifying all of the unit tests
with unroll bound 5. strpbrk_fail.c and strspn_fail.c both
require a higher unroll bound. The lowest bound that verifies
both of these is 13, so that is now in the config file

All other tests should work with unroll bound 5.