Add Fortran support to top.py
Involves a few hacks: first, use sed to
modify ll files to have the proper debug
info version. Second, compile the fortran
library first, so the smack.mod file is
available for compilation of test files.
Last, and this is less of a hack, replace
'main' with 'MAIN_' as an entry point.