Add an LLVM pass to normalize loops.
Specifically, if a phi node in a loop header gets its value from a block in the
loop body, and the source block branches when terminating, the generated Boogie
code incorrectly updates the variable corresponding to the phi node. To fix
this, a forwarding block is added between the source block and the header. The
generated Boogie code then updates the value for the phi node in the forwarding
block, and other destination blocks keep the non-updated value for the phi
node. Loops of this form are usually generated when compiling with
optimization. This fixes #314.