Make escape characters more robust
Extend Naming::escape's functionality
to escape many more characters.

Because Boogie can't handle backslashes in
identifiers, the call to llvm::DOT::EscapeString
has been removed. All of the characters that this
method currently escapes have been added to
Naming::escape

Most characters are replaced with underscores.
Theoretically, this could cause naming conflicts
in the future, but the probablity of two names
being identical except for their special characters
is low enough that it's worth it to neglect that
possibility right now.