Make escape characters more robust (#331)
* Make escape characters more robust

Extend Naming::escape's functionality
to escape many more characters.

Because Boogie can't handle backslashes in
identifiers, the call to llvm::DOT::EscapeString
has been removed. All of the characters that this
method currently escapes have been added to
Naming::escape

Most characters are replaced with underscores.
Theoretically, this could cause naming conflicts
in the future, but the probablity of two names
being identical except for their special characters
is low enough that it's worth it to neglect that
possibility right now.

* Stop escaping dollar signs

Replace escaping '$' with a comment.

When SMACK verifies contracts, it
generates LLVM IR with $ characters which
should not be replaced. The removal of '$'
means that if any actual LLVM code uses '$'
in a name, SMACK will not work, but this is
not very often if ever.

* Stop escaping tab/newline, start escaping space

Tab and newline characters are not likely to be
found in a name, while spaces are.

* Remove distinction of escape characters

Before, the code closely followed and explained
how it followed llvm::DOT::EscapeString. This
moves away from that, instead we just escape
all of the characters we're interested in.

* Start escaping dashes

These are used by kotlin in names