Colorize Clang's output
When the input C program is not syntactically correct, SMACK redirects
Clang's output to stdout. However, the message is not colorized, which
can make debugging cumbersome as it's hard to quickly identify the errors.

This commit solves this problem by adding a special flag to Clang such that
the output is colorized if SMACK's output is a tty device, otherwise it is
disabled to avoid annoying color information printed in a file.