Remove calls to `gep_type_iterator::operator*`
The dereference operator for this iterator was removed
[here](https://reviews.llvm.org/D26594) but has no direct replacement.

Intermediate types in a `GetElementPointer` instruction are (to my
knowledge) either structs, pointers, or arrays. To check for a struct,
we use `getStructTypeOrNull`. If the type is not a struct, we call
`getIndexedType` to get the element of the pointer or array. Arrays can
be differentiated from pointers by calling `isBoundedSequential`.

The big changes in DSA/Local are mostly indentation, use `git diff -b`
to view the significant ones.