Remove alignment param from LLVM intrinsics
The `memset` and `memcpy` intrinsics no longer require an alignment
parameter as they take pointers to `i8`. This broke the old code which
got arguments to these intrinsics by index.

Additionally, we hardcode the volatility of the access into the
resulting Boogie code instead of generating an expression.

FIXME: Is `length` supposed to be in bytes or as multiples of the
underlying type?