Overhauled Prelude.cpp
Specifically, this commit does the following,

1. Restructure/refactor Prelude.cpp such that it is written in OO as opposed to
mimicking the old macros defined in smack.c. Now Boogie declarations are
generated using a few generator classes where the generation methods in some of
these classes such as IntOpGen generate declarations according to a look-up
table summarizing the declarations.

2. Move the remaining declarations from smack.c to Prelude.cpp such as
floating-point related declarations.

3. Make declaration generation sensitive to SMACK flags such as
`--bit-precise`. Namely, when it is not enabled, bit-vector related declarations
will not be generated.

4. Bug fixes such the floating-point fma operation.