fixed `-Wstring-conversion` warning reported by Apple's Clang
The signatures of the relevant functions contain two default arguments in a row:
`bool ShortNames = false, const Twine &Title = "",`. Clang seems unable to match
the arguments passed to these functions by their types. The fix removes the
string literal argument instead of adding `false` to the arguments because we
want the `Title` argument to be "", leading the titles in the dot files to be
the functions analyzed rather than the string literal.