Allow optional generation of Boogie assumes from llvm.assume intrinsics
LLVM has the llvm.assume intrinsic which is intended to allow a
compiler to communicate facts about a program to LLVM's
optimizers. These facts may be useful for verification, so this allows
them to be used. Additionally, checking of assumptions is possible in
order to diagnose faulty assumptions arising from source code or
compiler bugs.

Specifically, a new flag "--llvm-assumes" is added with three modes:
- "none": This ignores the llvm.assume intrinsic and is the default
  mode.
- "use": This enables generation of Boogie assumes from llvm.assume
  instructions.
- "check": This enables assertion checking of the validity of assumed
  statements.