Fixed over-collapsing of regions when malloc is encountered
In order to establish whether a pointer returned from an external
function can alias program-allocated memory, we used to generate
regions for return values of pointer types of all functions. We
used to do that no matter how the pointers would end up being used.
In the case of malloc, this would almost always lead to collapsing
of regions since malloc returns a pointer of type i8, but is
almost always cast into the appropriate type and then used.
This fix removes this behavior and avoids collapsing of regions
due to malloc.

Fixes #473