Moved gcd test to the failing test folder
The new Boogie VC encoding causes Z3 to not like mod and div
operators over integers when Corral is used. Hence, I had to
move this regression to the failing ones. We should get back
to it later.
I also removed a related regression that was almost certainly
added by accident while back.