Better error traces
This commit changes the error trace generation flow. Originally, we parse the
textual error trace and converts it into a more readable format. This commit
first converts the texual error trace into JSON and further pretty-print the
error trace. This enables us to have a canonical representation of the error
trace.

Pretty-printing includes removal of duplicate lines as well as Boogie assertion
failures that should be opaque to the users.

This commit also fixes #617