AnnotateLoopExits from Module to Function Pass
Changes AnnotateLoopExits from a ModulePass
to a FunctionPass.

With the previous ModulePass version, there were
some crashes happening in the regressions,
which seem to be because LoopInfoWrapperPass
doesn't work with ModulePass in the legacy pass
manager:
https://groups.google.com/g/llvm-dev/c/lVc2-0d2cRs/m/uwuOA_pXAgAJ

I've changed the pass to be a FunctionPass,
and now the regressions work.

This also means that the pass now stores
the "__SMACK_loop_exit" function as a
private member.