Unified the conversion from APInt to std::string
Starting from LLVM 13, there isn't a `toString` method for APInt
that returns a std::string. Instead, it forces users to pass in
a small vector of chars to hold the results. So I change all the
old `toString` calls to a sequence of statements where a
`SmallString` is created ,passed to new `to*String` calls, and
finally gets converted to std::string using explicit conversions.